requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.17
pydantic==2.4.2
python-dotenv==1.0.0
openpyxl==3.1.5
pandas==2.2.3
python-jose==3.3.0
passlib==1.7.4